[refactor] add assignor-form and remove unused code.
