+++ api-payable/src/integrations/payables/payable.controller.ts
@@ -50,7 +50,6 @@ export class PayableController {
   @Get()
   async getPage(@Query() queryDto: PayablePaginationDto) {
     const { page, limit, cursorId, includeAssignor, selectKeys } = queryDto;
-    console.log(`[Log:queryDto]:`, queryDto);
     const pagination = await this.payableService.getPage({
       page,
       limit,
